moved some functions out from Vector3
