+++ BAVCL/Geometric/Vector3/Fill.cs
@@ -1,3 +1,4 @@
+using System.Linq;
 using BAVCL.Core;
 
 namespace BAVCL.Geometric;
