+++ BAVCL/Geometric/Vector3/Vector3.cs
@@ -11,40 +11,9 @@ namespace BAVCL.Geometric
 		#endregion
 
 		// CONSTRUCTOR
-		public Vector3(GPU gpu, float[] value, bool cache = true) : 
-			base(gpu, value, 3, cache) {}
+		public Vector3(GPU gpu, float[] value, bool cache = true) : base(gpu, value, 3, cache) {}
 
-		public Vector3(GPU gpu, int length) : 
-			base(gpu, length, 3) {}
-
-
-		// Create Vector3
-		public static Vector3 Fill(GPU gpu, float Value, int Length)
-		{
-			return new Vector3(gpu, Enumerable.Repeat(Value, Length).ToArray());
-		}
-		public static Vector3 Zeros(GPU gpu, int Length)
-		{
-			return new Vector3(gpu, new float[Length]);
-		}
-
-
-		// DATA Management
-		public Vector3 Copy()
-		{
-			if (_id != 0)
-			{
-				return new Vector3(gpu, Pull());
-			}
-			return new Vector3(gpu, Value[..]);
-		}
-
-		public static Vector3 AccessRow(Vector3 vector, int vert_row)
-		{
-			vector.SyncCPU();
-			return new Vector3(vector.gpu, vector.Value[(vert_row * 3)..((vert_row + 1) * 3)]);
-		}
-
+		public Vector3(GPU gpu, int length) : base(gpu, length, 3) {}
 
 		// CONVERT TO GENERIC VECTOR
 		/*
